Extract RedisSetUp
